EN6-173 Extract EditorToolbar into own function component file
